Add ModelHelper to specs, change user_rating, pet_rating columns to floats

+++ db/migrate/20180629161933_create_pettings.rb
@@ -6,8 +6,8 @@ class CreatePettings < ActiveRecord::Migration[5.2]
       t.string :name
       t.string :breed
       t.string :location
-      t.integer :pet_rating
+      t.float :pet_rating
       t.string :description
     end
-  end
+  end 
 end
